make first module
